pyproject.toml
[project]
name = "inv-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk[extensions]>=1.7.0",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain[anthropic,openai]>=0.3.26",
    "langgraph>=0.5.3",
    "litellm>=1.74.4",
    "llama-index>=0.12.49",
    # "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-llms-gemini>=0.5.0",
    "llama-index-llms-google-genai>=0.2.5",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-readers-file>=0.4.11",
    "openai>=1.97.0",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
